requirements.txt
python-dotenv==1.0.0
pytest==7.2.2
asyncio==3.4.3